requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-Migrate==4.0.5
Werkzeug==2.3.7
SQLAlchemy==2.0.21
undetected-chromedriver==3.5.4
selenium==4.15.2
beautifulsoup4==4.12.2
requests==2.31.0
fake-useragent==1.4.0
Pillow==10.0.1
imagehash==4.3.1
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
sendgrid==6.10.0
python-dateutil==2.8.2
difflib2==0.1.0
psycopg2-binary==2.9.7
gunicorn==21.2.0
# Render-specific optimizations
whitenoise==6.6.0
